Mostly fix image serialization

+++ NAPS2.Sdk/Images/Storage/FileStorage.cs
@@ -7,21 +7,31 @@ namespace NAPS2.Images.Storage
 {
     public class FileStorage : IStorage
     {
-        public FileStorage(string fullPath)
+        private readonly bool shared;
+
+        public FileStorage(string fullPath) : this(fullPath, false)
+        {
+        }
+
+        public FileStorage(string fullPath, bool shared)
         {
             FullPath = fullPath ?? throw new ArgumentNullException(nameof(fullPath));
+            this.shared = shared;
         }
 
         public string FullPath { get; }
 
         public void Dispose()
         {
-            try
+            if (!shared)
             {
-                File.Delete(FullPath);
-            }
-            catch (IOException)
-            {
+                try
+                {
+                    File.Delete(FullPath);
+                }
+                catch (IOException)
+                {
+                }
             }
         }
     }
